+++ pkg/connectors/icinga2/api.go
@@ -17,6 +17,7 @@ type Host struct {
 	MaxCheckAttempts    int      `json:"max_check_attempts"`
 	CheckAttempt        int      `json:"check_attempt"`
 	Groups              []string `json:"groups"`
+	NotesUrl            string   `json:"notes_url"`
 }
 
 type ServiceResponse struct {
@@ -39,6 +40,7 @@ type service struct {
 	MaxCheckAttempts    int         `json:"max_check_attempts"`
 	CheckAttempt        int         `json:"check_attempt"`
 	Groups              []string    `json:"groups"`
+	NotesUrl            string      `json:"notes_url"`
 }
 
 type checkResult struct {